requirements.txt
packaging==23.2
Pillow==10.1.0
pytesseract==0.3.10
streamlit==1.28.0
streamlit-chat==0.1.1
python-dotenv==1.0.0
